+++ packages/service-core/src/index.ts
@@ -2,5 +2,6 @@ export * from "./auth.js";
 export * from "./config.js";
 export * from "./errors.js";
 export * from "./http.js";
+export * from "./metrics.js";
 export * from "./notifications.js";
 export * from "./shutdown.js";